Unite RefName and Reference private data
This changeset actually takes an interessting approach in uniting the
both much related classes: A Reference is always a RefName, but a
RefName must not always be a Reference. This is actually inheritance.
But we don't inherit - at least not in the public class. From a consumer
point of view, these are totally separated classes...

Close #43